+++ linux/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   dynamic_color
   file_selector_linux
   flutter_secure_storage_linux
+  media_kit_libs_linux
   sqlite3_flutter_libs
 )